Moved docs to jekyll
This commit converts all old dokuwiki pages to markdown and allows to
generate a html version via markdown.

Since most stuff of this was autogenerated there might be some silly
escaping, but I think this can be fixed if the pages are updated by
hand.

Due to a bug in the markdown parser (maruku) markdown elements inside of
html tags need to be double escaped for now. (see:
https://github.com/bhollis/maruku/issues/97).

I fixed all the dead links along (and removed links to not created wiki
pages).